feat: Ranking -> Post로 병합
