--- back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/dto/RankingRequest.java
@@ -1,17 +1,0 @@
-package wooteco.team.ittabi.legenoaroundhere.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Setter;
-import lombok.ToString;
-import wooteco.team.ittabi.legenoaroundhere.domain.PostSearch;
-
-@AllArgsConstructor
-@Setter
-@EqualsAndHashCode
-@ToString
-public class RankingRequest {
-
-    private String areaIds;
-    private String sectorIds;
-}
