--- back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/dto/RankingResponse.java
@@ -1,5 +1,0 @@
-package wooteco.team.ittabi.legenoaroundhere.dto;
-
-public class RankingResponse {
-
-}
